Moved config validation to StartupEventListeners
